remove the Gong temporal client and move the config to config.ts
